Implement till Spring authentication and authority.

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -1,12 +1,15 @@
 package net.engineeringdigest.journalApp.service;
 
 import net.engineeringdigest.journalApp.entity.JournalEntry;
+import net.engineeringdigest.journalApp.entity.User;
 import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
+import net.engineeringdigest.journalApp.repository.UserRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,9 +18,33 @@ public class JournalEntryService
 {
     @Autowired
     JournalEntryRepository journalEntryRepository;
+    @Autowired
+    UserService userService;
+    @Autowired
+    private UserRepository userRepository;
+
+    @Transactional
+    public void saveEntry(JournalEntry journalEntry, String userName)
+    {
+        try {
+            User user = userService.findByUserName(userName);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            //user.setUserName(null);
+            userService.saveUser(user);
+
+        }catch (Exception e)
+        {
+            System.out.println("An error occured while saving the entry"+e);
+            throw new RuntimeException("An error occured while saving the entry", e);
+        }
+
+    }
 
     public void saveEntry(JournalEntry journalEntry)
     {
+
         journalEntryRepository.save(journalEntry);
     }
     public List<JournalEntry> getAll()
@@ -29,10 +56,25 @@ public class JournalEntryService
        return journalEntryRepository.findById(id);
     }
 
-    @DeleteMapping
-    public void deleteById(ObjectId id)
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName)
     {
-        journalEntryRepository.deleteById(id);
+        boolean removed = false;
+        try {
+
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }
+        catch (Exception e)
+        {
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the deleting", e);
+        }
+        return removed;
 
     }
 }
